requirements.txt
Pillow>=8.0.0
exifread>=2.3.0